// src/app/models/clasa.model.ts
export class Clasa {
  id: number;
  cifra: string;
  litera : string;
  total: string;
  romi: string;
  dizabilitati: string;
  parinti: string;
  burse: string;
  repetenti: string;
  online: string;
  remediala: string;
  total2b: string;
  romi2b: string;
  dizabilitati2b: string;
  parinti2b: string;
  burse2b: string;
  repetenti2b: string;
  online2b: string;
  remediala2b: string;
  cladire: string;
  predare: string;
  mixt: boolean;
  ani: string;
  data: string;
  unitateId: number;
  burse2bC: string;
  burseC: string;
  cifraC: string;
  cladireC: string;
  dataC: string;
  dizabilitati2bC: string;
  dizabilitatiC: string;
  literaC: string;
  mixtC: boolean;
  moderator: string;
  online2bC: string;
  onlineC: string;
  parinti2bC: string;
  parintiC: string;
  predareC: string;
  remediala2bC: string;
  remedialaC: string;
  repetenti2bC: string;
  repetentiC: string;
  romi2bC: string;
  romiC: string;
  romiE: string;
  total2bC: string;
  totalC: string;
  aniC: string;

  constructor(input?: any) {
    Object.assign(this, input);
  }
}